[RENAME] #47 - SuccessMessage로 변경
