+++ src/main/java/com/soptie/server/doll/message/SuccessMessage.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Getter
-public enum ResponseMessage {
+public enum SuccessMessage {
 	SUCCESS_GET_IMAGE("인형 이미지 조회 성공"),
 	;
