+++ src/main/java/com/soptie/server/member/controller/MemberController.java
@@ -15,9 +15,9 @@ import java.net.URI;
 import java.security.Principal;
 
 import static com.soptie.server.common.dto.Response.success;
-import static com.soptie.server.member.message.ResponseMessage.SUCCESS_CREATE_PROFILE;
-import static com.soptie.server.member.message.ResponseMessage.SUCCESS_GIVE_COTTON;
-import static com.soptie.server.member.message.ResponseMessage.*;
+import static com.soptie.server.member.message.SuccessMessage.SUCCESS_CREATE_PROFILE;
+import static com.soptie.server.member.message.SuccessMessage.SUCCESS_GIVE_COTTON;
+import static com.soptie.server.member.message.SuccessMessage.*;
 
 @RestController
 @RequiredArgsConstructor
