+++ src/main/java/com/soptie/server/routine/message/SuccessMessage.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Getter
-public enum ResponseMessage {
+public enum SuccessMessage {
 	SUCCESS_GET_THEME("데일리 루틴 테마 조회 성공"),
 	SUCCESS_GET_ROUTINE("데일리 루틴 조회 성공"),
 	SUCCESS_GET_HAPPINESS_THEME("행복 루틴 테마 조회 성공"),
