+++ src/test/java/com/soptie/server/member/controller/MemberControllerTest.java
@@ -5,14 +5,9 @@ import com.soptie.server.base.BaseControllerTest;
 import com.soptie.server.common.dto.Response;
 import com.soptie.server.member.dto.CottonCountResponse;
 import com.soptie.server.member.dto.MemberProfileRequest;
-import com.soptie.server.member.entity.CottonType;
-import com.soptie.server.routine.dto.DailyRoutinesResponse;
-import com.soptie.server.routine.fixture.DailyRoutineFixture;
 import com.soptie.server.doll.entity.Doll;
 import com.soptie.server.doll.entity.DollImage;
 import com.soptie.server.member.dto.MemberHomeInfoResponse;
-import com.soptie.server.member.dto.MemberProfileRequest;
-import com.soptie.server.member.entity.Cotton;
 import com.soptie.server.member.entity.Member;
 import com.soptie.server.member.entity.SocialType;
 import com.soptie.server.memberDoll.entity.MemberDoll;
@@ -32,9 +27,9 @@ import static com.epages.restdocs.apispec.ResourceDocumentation.resource;
 import static com.soptie.server.common.dto.Response.success;
 import static com.soptie.server.doll.entity.DollType.BROWN;
 import static com.soptie.server.member.entity.CottonType.DAILY;
-import static com.soptie.server.member.message.ResponseMessage.SUCCESS_CREATE_PROFILE;
-import static com.soptie.server.member.message.ResponseMessage.SUCCESS_GIVE_COTTON;
-import static com.soptie.server.member.message.ResponseMessage.*;
+import static com.soptie.server.member.message.SuccessMessage.SUCCESS_CREATE_PROFILE;
+import static com.soptie.server.member.message.SuccessMessage.SUCCESS_GIVE_COTTON;
+import static com.soptie.server.member.message.SuccessMessage.*;
 import static org.mockito.Mockito.when;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.*;
